% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgconfig.R
\name{pkgconfig}
\alias{pkgconfig}
\title{Configure compilation flags}
\usage{
pkgconfig()
}
\value{
String containing flags to add to \code{PKG_LIBS} of the \code{Makevars}.
}
\description{
Configure flags for compiling downstream packages.
}
\examples{
pkgconfig()

}
\author{
Aaron Lun
}
